+++ node/network/approval-distribution/src/lib.rs
@@ -23,7 +23,7 @@
 use futures::{channel::oneshot, FutureExt as _};
 use polkadot_node_network_protocol::{
 	grid_topology::{
-		AggressionConfig, RandomRouting, RequiredRouting, SessionGridTopologies,
+		RandomRouting, RequiredRouting, SessionGridTopologies,
 		SessionGridTopology,
 	},
 	v1 as protocol_v1, PeerId, UnifiedReputationChange as Rep, View,
@@ -90,6 +90,60 @@ impl RecentlyOutdated {
 
 	fn is_recent_outdated(&self, hash: &Hash) -> bool {
 		self.buf.contains(hash)
+	}
+}
+
+// In case the original gtid topology mechanisms don't work on their own, we need to trade bandwidth
+// for protocol liveliness by introducing aggression.
+//
+// Aggression has 3 levels:
+//
+//  * Aggression Level 0: The basic behaviors described above.
+//  * Aggression Level 1: The originator of a message sends to all peers. Other peers follow the rules above.
+//  * Aggression Level 2: All peers send all messages to all their row and column neighbors.
+//    This means that each validator will, on average, receive each message approximately `2*sqrt(n)` times.
+// The aggression level of messages pertaining to a block increases when that block is unfinalized and
+// is a child of the finalized block.
+// This means that only one block at a time has its messages propagated with aggression > 0.
+//
+// A note on aggression thresholds: changes in propagation apply only to blocks which are the
+// _direct descendants_ of the finalized block which are older than the given threshold,
+// not to all blocks older than the threshold. Most likely, a few assignments struggle to
+// be propagated in a single block and this holds up all of its descendants blocks.
+// Accordingly, we only step on the gas for the block which is most obviously holding up finality.
+
+/// Aggression configuration representation
+#[derive(Clone)]
+struct AggressionConfig {
+	/// Aggression level 1: all validators send all their own messages to all peers.
+	l1_threshold: Option<BlockNumber>,
+	/// Aggression level 2: level 1 + all validators send all messages to all peers in the X and Y dimensions.
+	l2_threshold: Option<BlockNumber>,
+	/// How often to re-send messages to all targeted recipients.
+	/// This applies to all unfinalized blocks.
+	resend_unfinalized_period: Option<BlockNumber>,
+}
+
+impl AggressionConfig {
+	/// Returns `true` if block is not too old depending on the aggression level
+	fn is_age_relevant(&self, block_age: BlockNumber) -> bool {
+		if let Some(t) = self.l1_threshold {
+			block_age >= t
+		} else if let Some(t) = self.resend_unfinalized_period {
+			block_age > 0 && block_age % t == 0
+		} else {
+			false
+		}
+	}
+}
+
+impl Default for AggressionConfig {
+	fn default() -> Self {
+		AggressionConfig {
+			l1_threshold: Some(10),
+			l2_threshold: Some(25),
+			resend_unfinalized_period: Some(5),
+		}
 	}
 }
